requirements.txt
django==3.1.1
django-extensions==3.0.7
django-oauth-toolkit==1.3.2
djangorestframework==3.11.1
